+++ tests/utils.py
@@ -1,0 +1,3 @@
+########################################################################
+def match_all_elements(eid):
+    return True
